backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.1
bcrypt==4.1.2
Werkzeug==3.0.1
gunicorn==21.2.0
